modification addPerson, addBook, creation des methodes(removeBook,removePerson, editeBook, EditePerson ,displaybook,displayPerson) et changement du type de l'attribut RentaPrice en double ????

+++ src/DaoLibrary/Book.java
@@ -94,7 +94,7 @@ public class Book {
 	}
 
 
-	public void setRentalPrice(int rentalPrice) {
+	public void setRentalPrice(double rentalPrice) {
 		this.rentalPrice = rentalPrice;
 	}
